Office AddIn - Excel Module - FY21Q2 refresh
- update demos to reflect latest dependencies
