+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/01 Tables and Charts/package.json
@@ -28,29 +28,29 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.10.2",
-    "@babel/polyfill": "^7.10.1",
-    "@babel/preset-env": "^7.10.2",
+    "@babel/core": "^7.11.6",
+    "@babel/polyfill": "^7.11.5",
+    "@babel/preset-env": "^7.11.5",
     "@types/find-process": "1.2.0",
     "@types/office-js": "^1.0.108",
     "@types/office-runtime": "^1.0.14",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "eslint-config-office-addins": "^1.0.15",
+    "copy-webpack-plugin": "^6.1.1",
+    "eslint-config-office-addins": "^1.0.19",
     "find-process": "^1.4.3",
     "file-loader": "^4.2.0",
     "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^4.3.0",
-    "office-addin-cli": "^1.0.10",
-    "office-addin-debugging": "^3.0.31",
-    "office-addin-dev-certs": "^1.5.2",
-    "office-addin-lint": "^1.0.23",
-    "office-addin-manifest": "1.5.0",
+    "html-webpack-plugin": "^4.5.0",
+    "office-addin-cli": "^1.0.13",
+    "office-addin-debugging": "^3.0.34",
+    "office-addin-dev-certs": "^1.5.5",
+    "office-addin-lint": "^1.0.26",
+    "office-addin-manifest": "1.5.7",
     "office-addin-prettier-config": "^1.0.12",
     "source-map-loader": "^0.2.4",
     "ts-loader": "^6.2.2",
-    "typescript": "^3.9.5",
+    "typescript": "^4.0.3",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
