+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/01 Tables and Charts/src/taskpane/taskpane.html
@@ -33,6 +33,7 @@
       <button class="ms-Button" id="filter-table">Filter Table</button><br/><br/>
       <button class="ms-Button" id="sort-table">Sort Table</button><br/><br/>
       <button class="ms-Button" id="create-chart">Create Chart</button><br/><br/>
+
     </main>
 </body>
