+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/02 Worksheets and Add-in Commands/src/taskpane/taskpane.html
@@ -34,6 +34,7 @@
       <button class="ms-Button" id="sort-table">Sort Table</button><br/><br/>
       <button class="ms-Button" id="create-chart">Create Chart</button><br/><br/>
       <button class="ms-Button" id="freeze-header">Freeze Header</button><br/><br/>
+
     </main>
 </body>
