+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/03 Dialogs/src/dialogs/popup.js
@@ -4,7 +4,6 @@
     .then(function () {
 
       document.getElementById("ok-button").onclick = sendStringToParentPage;
-
     });
 
   function sendStringToParentPage() {
