+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/03 Dialogs/src/taskpane/taskpane.html
@@ -36,6 +36,7 @@
       <button class="ms-Button" id="freeze-header">Freeze Header</button><br/><br/>
       <button class="ms-Button" id="open-dialog">Open Dialog</button><br/><br/>
       <label id="user-name"></label><br/><br/>
+
     </main>
 </body>
