+++ OfficeAddin/02 Building Add-ins for Microsoft Excel/Demos/03 Dialogs/src/taskpane/taskpane.js
@@ -3,6 +3,11 @@
  * See LICENSE in the project root for license information.
  */
 
+// images references in the manifest
+import "../../assets/icon-16.png";
+import "../../assets/icon-32.png";
+import "../../assets/icon-80.png";
+
 /* global console, document, Excel, Office */
 
 Office.onReady(info => {
@@ -144,34 +149,6 @@ function freezeHeader() {
     });
 }
 
-function toggleProtection(args) {
-  Excel.run(function (context) {
-    var sheet = context.workbook.worksheets.getActiveWorksheet();
-    sheet.load('protection/protected');
-
-    return context.sync()
-      .then(
-        function () {
-          if (sheet.protection.protected) {
-            sheet.protection.unprotect();
-          } else {
-            sheet.protection.protect();
-          }
-        }
-      )
-      .then(context.sync);
-  })
-    .catch(function (error) {
-      console.log("Error: " + error);
-      if (error instanceof OfficeExtension.Error) {
-        console.log("Debug info: " + JSON.stringify(error.debugInfo));
-      }
-    });
-  args.completed();
-}
-
-var dialog = null;
-
 function openDialog() {
   Office.context.ui.displayDialogAsync(
     'https://localhost:3000/popup.html',
@@ -188,3 +165,4 @@ function processMessage(arg) {
   document.getElementById("user-name").innerHTML = arg.message;
   dialog.close();
 }
+var dialog = null;
